moved components into folder
